Basic integration for commission and task page
