+++ server/frontend/src/pages/CommissionJson/Sample.json
@@ -1,0 +1,91 @@
+{
+  "$id": "143",
+  "$type": "MVX.Domain.Work.WorkTableDefinition, RioTinto.AutoHaul.Core",
+  "ID": "89a771fc-62bf-4785-94bc-47718f70856d",
+  "Title": "Visual Inspections & Auxiliary Checks",
+  "PreviousVersionTableID": "0a3e4f50-8d15-488b-8349-9922b79d486d",
+  "ManualCompleteAllowed": false,
+  "ShowCurrentProgress": true,
+  "ApplyGlobalChanges": false,
+  "ReferenceName": "Table 3",
+  "Columns": [
+    {
+      "$id": "144",
+      "$type": "MVX.Domain.Work.ColumnDefinition, RioTinto.AutoHaul.Core",
+      "ID": 1,
+      "HeaderText": "Action",
+      "ColumnData": "",
+      "WidthWeight": 4,
+      "ColumnType": 0,
+      "IsOptional": true,
+      "IsHidden": false,
+      "ResetValueOnLocalChange": false,
+      "ForceResetValueOnGlobalChange": false,
+      "ResetValueOnGlobalChange": false,
+      "PostChangeResetAction": false
+    },
+    {
+      "$id": "145",
+      "$type": "MVX.Domain.Work.ColumnDefinition, RioTinto.AutoHaul.Core",
+      "ID": 2,
+      "HeaderText": "Expected Result",
+      "ColumnData": "",
+      "WidthWeight": 2,
+      "ColumnType": 0,
+      "IsOptional": true,
+      "IsHidden": false,
+      "ResetValueOnLocalChange": false,
+      "ForceResetValueOnGlobalChange": false,
+      "ResetValueOnGlobalChange": false,
+      "PostChangeResetAction": false
+    },
+    {
+      "$id": "146",
+      "$type": "MVX.Domain.Work.ColumnDefinition, RioTinto.AutoHaul.Core",
+      "ID": 3,
+      "HeaderText": "Outcome",
+      "ColumnData": "",
+      "WidthWeight": 1,
+      "ColumnType": 1,
+      "IsOptional": false,
+      "IsHidden": false,
+      "ResetValueOnLocalChange": false,
+      "ForceResetValueOnGlobalChange": false,
+      "ResetValueOnGlobalChange": false,
+      "PostChangeResetAction": false
+    },
+    {
+      "$id": "147",
+      "$type": "MVX.Domain.Work.ColumnDefinition, RioTinto.AutoHaul.Core",
+      "ID": 4,
+      "HeaderText": "Signature",
+      "ColumnData": "",
+      "WidthWeight": 1,
+      "ColumnType": 4,
+      "IsOptional": false,
+      "IsHidden": false,
+      "ResetValueOnLocalChange": false,
+      "ForceResetValueOnGlobalChange": false,
+      "ResetValueOnGlobalChange": false,
+      "PostChangeResetAction": false
+    },
+    {
+      "$id": "148",
+      "$type": "MVX.Domain.Work.ColumnDefinition, RioTinto.AutoHaul.Core",
+      "ID": 5,
+      "HeaderText": "Notes",
+      "ColumnData": "",
+      "WidthWeight": 1,
+      "ColumnType": 0,
+      "IsOptional": true,
+      "IsHidden": false,
+      "ResetValueOnLocalChange": false,
+      "ForceResetValueOnGlobalChange": false,
+      "ResetValueOnGlobalChange": false,
+      "PostChangeResetAction": false
+    }
+  ],
+  "Rows": [
+    
+  ]
+}
