--- server/frontend/src/pages/TaskJson/OneTask.json
@@ -1,51 +1,0 @@
-{
-  "$id": "4",
-  "$type": "MVX.Domain.Work.WorkStepDefinition, RioTinto.AutoHaul.Core",
-  "ItemNumber": "1.1",
-  "SupportingImages": [
-    {
-      "$id": "5",
-      "$type": "Torq.Domain.Software.Attachment, Torq.Domain",
-      "FileName": "Image01.jpg",
-      "Folder": {
-        "$id": "6",
-        "$type": "Torq.Domain.Software.AttachmentFolder, Torq.Domain",
-        "Attachments": [
-          {
-            "$ref": "5"
-          }
-        ],
-        "Host": {
-          "$ref": "4"
-        }
-      }
-    }
-  ],
-  "Instructions": "1. [FFFFFF]Inspect Locomotive entry points.[-]\r\n\r\n2. [FFFFFF]Ensure Roads are locked out and the Locomotive isolated. [-][0070C0](Locomotive can be moved or other rolling stock can enter work area)[-]\r\n\r\n3. [FFFFFF]Ensure that safe working protection is in place. [-]\r\n\r\n4. [FFFFFF]Refer to Isolation matrix for lockout points.[-]\r\n\r\n5. [FFFFFF]Ensure entry walkways are properly secured with straps, & barriers are across any gaps. [-][0070C0](Falling down to the pit can occur[-][0070C0])[-]\r\n\r\n",
-  "Discipline": {
-    "$id": "7",
-    "$type": "MVX.Domain.Work.WorkDiscipline, RioTinto.AutoHaul.Core",
-    "ID": "70853c34-7a90-4343-ae93-1537d85327ee",
-    "Title": "",
-    "Description": "",
-    "StylisedText": ""
-  },
-  "SceneViews": [],
-  "Components": "",
-  "Location": "Step 1 over pit area",
-  "ConversionErrorLog": "",
-  "Alerts": "[FFFF00][b]Ask your Permit Isolation Officer or Administrator if unsure of isolation.  [/b][-]\r\n\r\n[FFFF00][b]Trips and falls can occur whilst boarding Locomotive, unwanted movement of walkways can occur[/b][-]",
-  "Attachments": [],
-  "PreviousWorkStepDefinitionID": "45112aa5-31f2-4310-8083-bc5c3dc09972",
-  "ReferencedTables": [],
-  "Title": "Sign on to permit of isolation and lock on",
-  "ID": "e8a475b7-28f2-4a74-ae2b-6cfd09bc85d5",
-  "ParentItem": {
-    "$ref": "3"
-  },
-  "ParentID": "00000000-0000-0000-0000-000000000000",
-  "ParentFolder": null,
-  "AttachmentFolder": {
-    "$ref": "6"
-  }
-}
