+++ server/frontend/src/pages/TaskJson/task-tabs/instruction.css
@@ -1,9 +1,9 @@
-/* .instruction-container {
+.instruction-container {
   display: flex;
   align-items: center;
   justify-content: center;
 
-} */
+}
 
 
 .instruction-content {
